error handling for the rest
