--- src/services/mod.rs
@@ -1,15 +1,0 @@
-//! TODO.
-//!
-//!
-
-pub use api_keys::ApiKeys;
-pub use audiences::Audiences;
-pub use contacts::Contacts;
-pub use domains::Domains;
-pub use emails::Emails;
-
-mod api_keys;
-mod audiences;
-mod contacts;
-mod domains;
-mod emails;
