--- src/types/mod.rs
@@ -1,17 +1,0 @@
-//! TODO.
-//!
-//!
-
-pub use api_keys::*;
-pub use audiences::*;
-pub use contacts::*;
-pub use domains::*;
-pub use emails::*;
-pub use error::*;
-
-mod api_keys;
-mod audiences;
-mod contacts;
-mod domains;
-mod emails;
-mod error;
